noDataModal will clear filters and handleApply
