+++ src/Components/Filters/urlEncodeDecode/urlEncode.js
@@ -11,7 +11,6 @@ const createSearchParams = (filters, urlSearchParams) => {
         "filter" + String(filterId)
       ] = `${filters[filterId].selectedTableColumnName}`;
     }
-    console.log("options", options);
     if (Number(filterId) > 1 && filters[filterId].selectedTableColumnName) {
       urlSearchParams[
         "filter" + String(filterId)
@@ -24,11 +23,9 @@ const createSearchParams = (filters, urlSearchParams) => {
 
 const createUrl = urlSearchParams => {
   let keys = Object.keys(urlSearchParams);
-  console.log("urlSearchParams", urlSearchParams);
   let values = Object.values(urlSearchParams).map(value =>
     encodeURIComponent(value)
   );
-  console.log("values", values);
   const filterStrings = keys.map((key, i) => key + "=" + values[i]).join("&");
   return filterStrings;
 };
