+++ src/Components/reducers/catReducer.js
@@ -3,7 +3,6 @@ import { LABELS_SUCCESS } from "../redux-actions/catLabels";
 const initialState = {
   labels: {}
 };
-console.log("labels", initialState.labels);
 const catLabelReducer = (state = initialState, action) => {
   switch (action.type) {
     case LABELS_SUCCESS:
