+++ src/DataParseHelpers/filterByDate.js
@@ -3,7 +3,6 @@ function filterByDate(data, startDate, endDate) {
   startDate = startDate.replace(/-/g, "");
   endDate = endDate.replace(/-/g, "");
   const filteredData = data.filter(obj => {
-    //console.log('obj',obj)
     let objectDate = +obj.created_date.split("T")[0].replace(/-/g, "");
     if (objectDate > startDate && objectDate < endDate) {
       return obj;
